Busqueda filtrada de personajes

+++ src/main/java/com/alkemy/disney/disney/dto/PersonajeDTO.java
@@ -13,7 +13,7 @@ public class PersonajeDTO{
     private Long id;
     private String imagen;
     private String nombre;
-    private int edad;
+    private Long edad;
     private double peso;
     private String historia;
     Set<PeliculaSinPersonajesDTO>  peliculas  = new HashSet<>();
